+++ src/spmv/SpMV_ELL_openmp.c
@@ -22,7 +22,7 @@ void ELLMatrix_SpMV_OPENMP(const ELLMatrix *matrix,const Vector *x, Vector *y, S
     }
     start = omp_get_wtime();
 #pragma omp parallel for schedule(auto) default(none) shared(matrix, x, y, stderr)
-    for (int row = 0; row < matrix->row_size; row++) {
+    for (u_int64_t row = 0; row < matrix->row_size; row++) {
         float dot = 0.0f;
         for (size_t i = 0; i < matrix->num_elem; i++) {
             size_t index = row * matrix->num_elem + i;
